requirements.txt
Django==4.1
groq==0.9.0
django-environ==0.11.2
gunicorn==22.0.0
whitenoise==6.6.0
dj-database-url==2.2.0
psycopg2==2.9.9
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
regex==2024.5.15
requests==2.31.0
python-dotenv==1.0.1
pillow==10.3.0
parso==0.8.3
pdfminer.six==20240706

